+++ src/main/java/fly/be/flyflix/auth/controller/AuthController.java
@@ -3,7 +3,9 @@ package fly.be.flyflix.auth.controller;
 import fly.be.flyflix.auth.controller.dto.LoginRequest;
 import fly.be.flyflix.auth.controller.dto.LoginResponse;
 import fly.be.flyflix.auth.entity.Aluno;
+import fly.be.flyflix.auth.entity.Usuario;
 import fly.be.flyflix.auth.repository.AlunoRepository;
+import fly.be.flyflix.auth.repository.UsuarioRepository;
 import fly.be.flyflix.auth.service.EmailService;
 import fly.be.flyflix.auth.service.TokenService;
 import jakarta.validation.Valid;
@@ -19,16 +21,17 @@ import java.util.Optional;
 public class AuthController {
 
     private final TokenService tokenService;
-    private final AlunoRepository alunoRepository;
+    private final UsuarioRepository usuarioRepository;
     private final EmailService emailService;
     private final PasswordEncoder passwordEncoder;
 
     public AuthController(TokenService tokenService,
-                          AlunoRepository alunoRepository,
+                          UsuarioRepository usuarioRepository,
                           EmailService emailService,
                           PasswordEncoder passwordEncoder) {
         this.tokenService = tokenService;
-        this.alunoRepository = alunoRepository;
+        this.usuarioRepository = usuarioRepository;
+
         this.emailService = emailService;
         this.passwordEncoder = passwordEncoder;
     }
@@ -37,13 +40,13 @@ public class AuthController {
     public ResponseEntity<?> esqueciSenha(@RequestBody Map<String, String> body) {
         String email = body.get("email");
 
-        Optional<Aluno> alunoOpt = alunoRepository.findByEmail(email);
-        if (alunoOpt.isEmpty()) {
+        Optional<Usuario> usuarioOpt = usuarioRepository.findByEmail(email);
+        if (usuarioOpt.isEmpty()) {
             return ResponseEntity.badRequest().body("Email não cadastrado.");
         }
 
-        Aluno aluno = alunoOpt.get();
-        String token = tokenService.gerarTokenRedefinicaoSenha(aluno);
+        Usuario usuario = usuarioOpt.get();
+        String token = tokenService.gerarTokenRedefinicaoSenha(usuario);
 
         String link = "http://localhost:3000/resetar-senha?token=" + token;
         emailService.enviarEmail(email, "Redefinição de senha Flyflix",
@@ -52,30 +55,40 @@ public class AuthController {
         return ResponseEntity.ok("Email enviado para redefinição de senha.");
     }
 
+
+
     @PostMapping("/resetar-senha")
     public ResponseEntity<?> resetarSenha(@RequestBody Map<String, String> body) {
         String token = body.get("token");
         String novaSenha = body.get("novaSenha");
 
-        try {
-            Aluno aluno = tokenService.validarTokenRedefinicaoSenha(token);
-            aluno.setSenha(passwordEncoder.encode(novaSenha));
-            alunoRepository.save(aluno);
-            tokenService.invalidarToken(token);
+        // Validação da nova senha
+        if (!isSenhaValida(novaSenha)) {
+            return ResponseEntity.badRequest().body("A senha deve ter no mínimo 8 caracteres, uma letra maiúscula, um número e um caractere especial.");
+        }
 
+        try {
+            Usuario usuario = tokenService.validarTokenRedefinicaoSenha(token);
+            usuario.setSenha(passwordEncoder.encode(novaSenha));
+            usuarioRepository.save(usuario);
+            tokenService.invalidarToken(token);
             return ResponseEntity.ok("Senha alterada com sucesso.");
         } catch (RuntimeException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
+    private boolean isSenhaValida(String senha) {
+        String regex = "^(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+=<>?{}\\[\\]~]).{8,}$";
+        return senha != null && senha.matches(regex);
+    }
+
+
+
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody @Valid LoginRequest loginRequest) {
-        try {
-            LoginResponse response = tokenService.login(loginRequest);
-            return ResponseEntity.ok(response);
-        } catch (Exception e) {
-            return ResponseEntity.status(401).body("Credenciais inválidas.");
-        }
+        LoginResponse response = tokenService.login(loginRequest);
+        return ResponseEntity.ok(response);
     }
+
 }
