+++ src/main/java/fly/be/flyflix/auth/exception/EmailNotFoundException.java
@@ -1,0 +1,9 @@
+package fly.be.flyflix.auth.exception;
+
+public class EmailNotFoundException extends RuntimeException {
+    public EmailNotFoundException() {
+        super("E-mail inválido.");
+    }
+}
+
+
