+++ src/main/java/fly/be/flyflix/auth/exception/InvalidPasswordException.java
@@ -1,0 +1,7 @@
+package fly.be.flyflix.auth.exception;
+
+public class InvalidPasswordException extends RuntimeException {
+    public InvalidPasswordException() {
+        super("Senha inválida.");
+    }
+}
